requirements.txt
# Core
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20

# AI
google-genai==1.14.0

# File parsing
pdfplumber==0.11.6

# SSE
sse-starlette==2.2.1

# Dev / Test
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
